[Bug Fix]
1.Fix the bug of find products by level1_id and level2_id and page split bug.

+++ src/ca/mcmaster/shop/product/ProductDao.java
@@ -54,23 +54,13 @@ public class ProductDao extends HibernateDaoSupport {
 		return product;
 	}
 
-	/**
-	 * @param id
-	 * @return
-	 */
 	public Integer findNumByLevel2_id(Integer id) {
 		@SuppressWarnings("unchecked")
-		List<Integer> list = this.getHibernateTemplate().find(
+		List<Long> list = this.getHibernateTemplate().find(
 				"select count(*) from Product p, Level2 lvl where p.product_belonging = lvl and lvl.level2_id = ?", id);
 		return list.get(0).intValue();
 	}
 
-	/**
-	 * @param id
-	 * @param begin
-	 * @param limit
-	 * @return
-	 */
 	public List<Product> findLevel2ProductByPage(Integer id, Integer begin, int limit) {
 		@SuppressWarnings("unchecked")
 		List<Product> list = this.getHibernateTemplate()
